Move and use editorHeight extension
