+++ client/web/src/notebooks/blocks/markdown/NotebookMarkdownBlock.tsx
@@ -10,8 +10,8 @@ import classNames from 'classnames'
 import PencilIcon from 'mdi-react/PencilIcon'
 import PlayCircleOutlineIcon from 'mdi-react/PlayCircleOutlineIcon'
 
-import { changeListener, editorHeight } from '@sourcegraph/search-ui'
-import { useCodeMirror } from '@sourcegraph/shared/src/components/CodeMirrorEditor'
+import { changeListener } from '@sourcegraph/search-ui'
+import { useCodeMirror, editorHeight } from '@sourcegraph/shared/src/components/CodeMirrorEditor'
 import { Markdown } from '@sourcegraph/shared/src/components/Markdown'
 import { ThemeProps } from '@sourcegraph/shared/src/theme'
 import { Icon } from '@sourcegraph/wildcard'
