+++ client/web/src/notebooks/blocks/query/NotebookQueryBlock.tsx
@@ -15,9 +15,9 @@ import {
     CodeMirrorQueryInput,
     changeListener,
     createDefaultSuggestions,
-    editorHeight,
 } from '@sourcegraph/search-ui'
 import { ActionItemAction } from '@sourcegraph/shared/src/actions/ActionItem'
+import { editorHeight } from '@sourcegraph/shared/src/components/CodeMirrorEditor'
 import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { HoverContext } from '@sourcegraph/shared/src/hover/HoverOverlay.types'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
